Move viewer.html to a flask app
Move viewer.html and all it's supporting javascript to a simple
Flask app that also servers the rows and stops files as json.

Modify viewer.html to read the date to analyse from a HTML fragment
identifier and then use xhr requests to retrieve the relevant rows
and stops data.