Allow bus to return to start
Previously, 'at_stop' could only be entered from 'before'. This deals
with the situation such as at Cambourn where a travelling bus passes
its start stop a second time part way through its journey.

But it's a problem if we fail to detect arrival at the destination
becasue there's a good chance the bus will then returrn to the origin
and then back to the destination, resulting in a detected trip that's
three times longer than it should be.

For the time being, allow 'at_start' to be entered from any state
which will at least correctly reset this situation.