Rework extract_segments
* Rework extract_segments to emit not only the locations on segments
between the 'to' and 'from' stops that we are interested in, but also
all the locations on the other segments that don't represent travel
between those stops. This helps debugging the split between 'on route'
and 'off route' segments.

* Rewrite the state transition machine

* Rename internal states to 'off_route', 'at_start', and 'on_route'

* Add a new transition 'on_route' --> 'off_route' if there's a pause
of mote than a few minutes between position reports (because the bus
has probably gone off somewhere that isn't part of a service)

* Allow for separate departure and arrival thresholds